mydemo/requirement.txt
asgiref==3.3.4
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.11.0
psycopg2==2.8
pyspark==2.4.8
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
